I added a new constant, max_mag_correlation, to constants.F90.
This variables defines the maximum absolute value (or magnitude) that
a correlation is allowed to have.  Thus, all correlations must have a
value between -max_mag_correlation and +max_mag_correlation.  The value 
of max_mag_correlation is set to 0.99.  The new constant is found in 
advance_xm_wpxp_module.F90 and in clip_explicit.F90.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/hoc_v2.2_tuner/trunk@2883 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 4d8de78a5406c73d1b7e9acdc2b11c33a5662b69