Several changes to the scalars:
* Added if..then's for the passive scalars that are meant to emulate 
rt and thl, so that when a variable is a covariance (e.g. sclr'thl' when
the scalar is thl) it will be clipped accordingly.  This won't affect the CO2
case.

* Did a similar if..then for the rhs term from dp1. Also won't affect CO2

* Fixed an error in the clipping of sclr'^2, where sclrtol was passed as the
tolerance rather than sclrtol^2.


Ran findent on the file.

The results for GABLS2 now match up nicely for sclram == thlm, and almost
as nicely for rt, except for what appears to be roundoff in the final hours.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/hoc_v2.2_tuner/trunk@2962 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 10afeeb4e53f70d2f6a8f1bafc9c0c18cf823c6b