I added model flag l_hyper_dfsn to model_flags.F90.
I also changed the code in advance_wp2_wp3_mod.F90 to declare
the number of superdiagonals (nsup) and the number of
subdiagonals (nsub) based on the l_hyper_dfsn flag.  These
variables are now declared in advance_wp2_wp3 and then passed into
wp23_solve and wp23_lhs.  The result of the changes are an
exact binary match for all cases and files.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/hoc_v2.2_tuner/trunk@2996 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 47a64ed0e15bb6dca2eee1e8aec1af12732ff2ff