I slightly altered the formula for adjusting the value of nu_hd based on grid box size.  The old version was of the form:
nu_hd = nu_hd * ( deltaz / 40.0 ).

The new version is of the form:

nu_hd = nu_hd * ( deltaz / 40.0 )^2.

The squared-adjuster version works more effectively for the case of
4th-order hyper-diffusion.  The value of nu_hd must be very large to
have any effect on cases with large grid box sizes (such as ARM 97 or
LBA, which have an average deltaz of 216.73228 m. for the stretched 
grid).  However, the value of nu_hd must be decreased dramatically in
order to avoid having substantial effects on cases with small grid
box sizes (such as DYCOMS-II (RF01, RF02 DO, and RF02 DS), which has 
a deltaz of 10 m., and ATEX, which has a deltaz of 20 m.).  After a
lot of testing, I decided that 10,000 was a good value for nu_hd
(before being adjusted based on the case).  I altered
default_parameters.in and high_dfsn_parameters.in to reflect that
value.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/hoc_v2.2_tuner/trunk@3002 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 4879b2bd13242600a66e0c06fa4bba061d88ccf4