I officially turned on 4th-order hyper-diffusion in the SVN repository in model_flags.F90.  It is applied to w'^2 and w'^3.  In addition, I also changed the value of nu_hd from 10000 to 20000 in default_parameters.in.
Until recently, turning on hyper-diffusion caused some problems for some
of the base cases.  For example, in the FIRE case, w'^3 was significantly
increased, and w'^2 and rcm were significantly decreased as a result of 
turning on 4th-order hyper-diffusion for w'^2 and w'^3.  The problem 
worsened if the value of nu_hd was increased much over 10000.  For example,
ATEX developed a w'^2 "bubble" above cloud top.  However, without being
able to increase nu_hd, the ARM 97 profiles were not significantly
smoothed-out by hyper-diffusion.

However, after I made the recent change of applying an adjuster to rtm/thlm
if w'r_t'/w'th_l' were clipped, the whole dynamic changed.  The base cases
that were previously effected by hyper-diffusion now are acting much more
robust.  Turning on hyper-diffusion hardly effects the results of the base
cases.  Furthermore, I can now substantially increase the value of nu_hd
without causing problems for the base cases.  This allows the ARM 97 profiles
to be smoothed-out somewhat.  After some testing, I found that a nu_hd value
of 20000 works best for ARM 97 while maintaining the integrity of the other
cases.  Furthermore, the maximum wp2 contour over the course of the ARM 97 
run is down to 35 m^2/s^2, and the maximum wp3 contour is down to
+/- 150 m^3/s^3.  Those are the best results that I have seen.

As a side note, since the average ARM 97 grid spacing is over 200 m., the
value of nu_hd for ARM 97 is now over 400,000.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/hoc_v2.2_tuner/trunk@3026 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: df396920758669e153374f524b4ec69668f7a2ed