I added the weighted time-stepping scheme to the x'^2 / x'y' turbulent advection term.  I also added it to the ta budget term.  I tested the budgets for a few of the xp2/xpyp variables, and the budgets balance.
The results for the base set of cases basically remains the same.
Stability has been increased for ARM 97.  For ARM 97 (60 sec. time
step), the wp2 contours range from 0 to 13 m^2/s^2, and the wp3
contours range from -10 to 35 m^3/s^3.  Those are the smallest
values that I have seen yet for those variables for ARM 97.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/hoc_v2.2_tuner/trunk@3049 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: d5f605d3c32efd9b1743758275649d11f5d1fecd