After thoroughly testing the code, I have officially switched um and vm over to the use of implicit momentum surface fluxes.  I have also double checked to make sure that the um and vm budgets still balance (and they do). In fact, the budgets balance for either form of surface fluxes (either the implicit momentum surface fluxes or the generalized explicit surface fluxes).
This does not effect any of the results for most of the base set of cases.
However, MPACE A and LBA are slightly changed, and ARM 97 shows much higher
extrema for wp2 and wp3 over the entire course of the run.  The instability
in GABLS2 and COBRA is cured, as they both now run to completion with at
least a 60 minute time step.

I performed a run_standalone_all.bash for standard (1 minute), 5 minute,
10 minute, 15 minute, 20 minute, 30 minute, 45 minute, and 60 minute
time steps.  The results are as follows:

standard (1 minute):  All cases run to completion.
5 minute:   RICO and ARM 97 blow up.
10 minute:  Only RICO blows up.
15 minute:  RICO and June 25 blow up.
20 minute:  Only RICO blows up.
30 minute:  RICO, ARM, and ARM 97 blow up.
45 minute:  RICO, ARM 97, and MPACE-A blow up.
60 minute:  Only RICO blows up.

Odd.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/hoc_v2.2_tuner/trunk@3061 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: aca5d455bc28abc2eb18e50c67162bfe3e7d3da8