I changed the value of emin based on the following reasoning: if anisotropic TKE is enabled, em = (1/2) * ( up2 + vp2 + wp2 ); otherwise, em = (3/2) * wp2.  Since up2, vp2, and wp2 all have the same minimum threshold value of wtol_sqd, em cannot be less than (3/2) * wtol_sqd.  Thus, emin = (3/2) * wtol_sqd.
I then changed a couple of thresholds on em in advance_clubb_core,
so that the thresholds of em are changed back to emin, rather than
wtol_sqd.  I also made the same change (to the same equations) in
clubb_driver.

I noticed that a piece of code was commented out in advance_clubb_core,
but the same piece of code (for the exact same equation), was left
in play in clubb_driver (for initialization).  I commented out the
piece of code in clubb_driver in order to mirror the commented-out
code in advance_clubb_core.  The commented-out code contained the
following comment:
    ! Modification to damp noise in stable region
! Vince Larson commented out because it may prevent turbulence from
!    initiating in unstable regions.  7 Jul 2007

A plotgen showed that the results of the base cases only changed
a hairline in a couple of cases.




git-svn-id: file:///usr/local/src/svnroot/clubb_repos/hoc_v2.2_tuner/trunk@3062 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: c40ad0ed5af7bac28d2dd1f461bac8a03ede42e7