I applied the "over-implicit" weighted time step to w'x' (w'r_t', w'th_l', w'sclr') pressure term 1 (pr1).
The wprtp and wpthlp budgets still balance.  There
is not much change to the base set of cases.  A plotgen
shows that LBA and MPACE-A have slight changes.  The
ARM 97 case changes again.  The contours for w'^2 range
from 0 to 27 m^2/s^2; while the contours for w'^3 range
from -150 to 150 m^3/s^3.

I performed a run_standalone_all.bash for standard (1 minute),
5 minute, 10 minute, 15 minute, 20 minute, 30 minute,
45 minute, and 60 minute time steps.  The results are as follows:

standard (1 minute):  All cases run to completion.
5 minute:   All cases run to completion.
10 minute:  All cases run to completion.
15 minute:  All cases run to completion.
20 minute:  RICO and ARM 97 blow up.
30 minute:  Only RICO blows up.
45 minute:  RICO and ARM 97 blow up.
60 minute:  RICO and ARM 97 blow up.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/hoc_v2.2_tuner/trunk@3081 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 3b30301431f91e3e3b175828fa586ace1ceca532