I finally integrated the new subroutine that calculates the mean upwards and mean downwards values of vertical velocity, w, with the rest of the code.
I found that this new code does not change the results of the
benchmark cases at the normal one minute time step.  Only ARM 97
changes slightly on the plots.  This also does not change the
stability of the code at longer time steps.  RICO is still the
only case to blow up, and it does so at 20 min., 30 min., 45 min.,
and 60 min. time steps, just as before.

I have officially flipped the flag to use this scheme to calculate
the number of "reachable" grid levels, rather than the constant
height scheme.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/hoc_v2.2_tuner/trunk@3185 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: d85ff2830da5c2a405ecf18806101edbfed418af