I added the new statistical parameters rrainm_src_adj and Nrm_src_adj to keep track of the amount of adjustment to the rain source terms that needs to be made when cloud water is over-depleted.
As a side note, the rrainm_src_adj and Nrm_src_adj terms do
not come into play in the DYCOMS-II RF02 (DO and DS) and RICO
cases when they are run at the standard one minute time step.
The binary differences in the GrADS files for those cases
(which I noted in my previous SVN commitment) was due to
addition being performed in a different order in the new coding.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/hoc_v2.2_tuner/trunk@3192 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 267035a406015ff969f3151e7e33bf7812e5b09a