I removed the l_kk_rain, l_coamps_micro, etc. flags and replaced them with a single variable, "micro_scheme" which contains a string for the microphysics case, e.g. "khairoutdinov_kogan" for l_kk_rain.
Added some logical variables for the Morrison scheme,
of which l_ice_micro and l_graupel are shared with coamps microphyics.

Removed the initialization of Ncm and Ncnm out of the tndcy subroutines.

The only case that changes was ARM 97, for unknown reasons.  The case
setup appears to be equivalent in every way (l_graupel off, l_ice_micro on)
but for some reason the results diverge.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@3280 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 2968a9f24dbdc03bd404f916eb279033aa661b23