I removed some of the case specific radiation code from the clubb_driver module.  However, some of the cases combined large-scale tendencies with analytic radiation in a way that doesn't appear to be seperable.  For these cases we might want to add an error or a warning that radiation is being double counted when BUGSrad is enabled.
The next step is to remove calls to rad_lwsw/cloud_rad from the individual
cases altogether, but this is some what tricky because although the
calculation of LW appears to be derived from the same formula for DYCOMS II 
RF01 & RF02, ATEX & BOMEX, and rad_lwsw, the code is actually somewhat
different.  Furthermore, Nov 11 has the option to run with a fixed value of
cosine of the solar zenith angle that isn't in any of the other altocu cases.

Note that MPACE A and MPACE B change slightly with the "t_since_solar_noon"
ad hoc formula removed.  The old formula contained some approximations with
relatively few significant digits (e.g. using 3.14 for pi) that accounts
for this.


git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@3403 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: a8d3b11de32fda4442a7e5e8fceaafd8f00971ed