I have written a code that handles mean advection of a thermodynamic-level variable at thermodynamic grid level k = gr%nnzp according to a constant derivative method (a.k.a. "one-sided" derivative method).
With this commitment, the mean fields set in advance_xm_wpxp_module.F90
now loop over the top level, just like they do for all other vertical levels
except for the k = 1 level (which is below the model surface and does not
effect anything).  This effects thlm, rtm, and any passive scalar fields.
The effects of vertical advection (or large-scale subsidence), any forcings,
and any turbulent advection or now included.  Since the monotonic flux
limiter is applied to the fields that are set in advance_xm_wpxp_module.F90,
it also needed to be modified in a similar fashion.

This code has been tested a lot by me, it is working properly.  This also
happens to take care of the thlm and rtm spikes at the upper boundary of
the GABLS3 night case.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@3993 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: bdfc32f4c9729cf5bea1c45434e6f7b06853798a