I committed some test code (commented out currently) for subroutine hydrostatic.
The idea of the subroutine is to solve for exner by the following
equation:

d(exner)/dz = - grav / ( Cp * thvm ).

In order to solve for exner, integration is needed:

INT(exner_1:exner_2) d(exner)
= - ( grav / Cp ) INT(z_1:z_2) (1/thvm) dz.

Currently, the code treats thvm as a constant over the layer in order
to pull it out of the integral, and then solves.

The new function calc_exner_const_theta solves for this method.

However, a better method might be to consider thvm as variable that
varies linearly between two successive grid levels.  Thus, within the
integral, thvm(z) = A*z + B, where A and B are constants.  This is
analagous to previous changes made in the calculation of Lscale, where
we treated d(CAPE)/dz as varying linearly between two successive levels,
rather than constant over a level.

The new function calc_exner_linear_theta solves for this new method.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@4300 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: d9dccce52825c3817f9dc357daf99fb8bb53865a