I am committing the new formulation for hydrostatic.
The main idea is to really nail down the computations of hydrostatic
exner, pressure, and density.  With the upcoming switch to the
anelastic set of equations, the calculation of dry, static, base-state
density becomes much more important.

Just as before, using theta as input into subroutine hydrostatic
will yield a good approximation for dry exner, pressure, and density,
while using theta_v as input into subroutine hydrostatic will yield
a good approximation for total (moisture included) exner, pressure,
and density.

Modifications have been made in order to provide for a better
calculation of exner, pressure, and density:  (1) at the interior
momentum levels (since rho_zm is output and since rho_ds_zm is a
very important variable in the new anelastic code); and (2) when
the grid levels are very far apart for deep convective cases,
and therefore linear interpolation of exner, pressure, and density
will not yield precise results (for the profiles of all three
do not change linearly over a large model domain, but rather change
more logarithmically with height).

Subroutine hydrostatic integrates the equation:

d(exner)/dz = - grav / ( Cp * th_var);

such that:

INT(exner_1:exner_2) d(exner)
= - ( grav / Cp ) INT(z_1:z_2) (1/th_var) dz;

where th_var is the theta variable (theta for a dry, hydrostatic
base-state, and thvm for the moist code).

The previous version simply figured th_var to be constant over the
depth of a level, thereby handling the integration by pulling the
"constant" th_var out of the integral.

The new version allows for the variation of th_var with height by
considering th_var as a variable that varies linearly with respect
to height.  Thus, th_var in the above integration is replaced by
an expression of the form A*z + B.  This integral is solved and
allows for a better calculation of exner (and thus pressure and
density) with respect to height.

This change is similar in form to a previous change made to the
calculation of Lscale.

A plotgen shows that all cases overlap, with the exception of
the sensitive cases ARM 97 and MPACE-A.  It should be pointed out
that TWP-ICE and the three cloud feedback cases do not use
subroutine hydrostatic (for their soundings are given in pressure
rather than height), so this change cannot effect them.

An interesting test might be to run the super-coarse deep convective
grid (30 level; approx. 34000 m.) and compare the difference with
the cases before this change.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@4302 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: bf5f3d32656ca9d58a41a58d9763c5ae74b4c1c3