I did some serious studying and deriving in regards to theta_v yesterday ...
I though that, while an approximated and linearized version of theta_v
works best for CLUBB core (for moments including thv' in pdf_closure,
theta_v of a parcel in compute_length, and a quick calculation of thvm
in advance_clubb_core), it would be best to use the most accurate version
possible in the initialization portion of CLUBB.  This is important because
it effects the setting of pressure, exner, and density, and thus also the
setting of dry, static, base-state density.

My summary is best stated in the comments I put in the code:

! Now, compute initial thvm, given initial thm, rtm, and rcm.
!
! The exact form of the equaton for theta_v (with only water vapor
! included) is:
!
! theta_v = theta * [ ( 1 + (R_v/R_d)*r_v ) / ( 1 + r_v ) ];
!
! which can be rearranged as:
!
! theta_v = theta * [ 1 + { (R_v/R_d) - 1 } * { r_v / ( 1 + r_v ) } ].
!
! The exact form of the equation for theta_v (including water vapor and
! cloud water) is:
!
! theta_v = theta * [ ( 1 + (R_v/R_d)*r_v ) / ( 1 + r_v + r_c ) ];
!
! which can be rearranged as:
!
! theta_v = theta * [ 1 + { (R_v/R_d) - 1 } * { r_v / ( 1 + r_v + r_c ) }
!                     - { r_c / ( 1 + r_v + r_c ) } ].
!
! This version is written with r_t = r_v + r_c, such that:
!
! theta_v = theta * [ 1 + { (R_v/R_d) - 1 }
!                         * { ( r_t - r_c ) / ( 1 + r_t ) }
!                     - { r_c / ( 1 + r_t ) } ].
!
! To use theta_l instead of theta, simply substitute the following for
! theta in the above expression:  theta_l + {L_v/(C_p*exner)} * r_c.
!
! The CLUBB code generally uses an approximated and linearized version of
! the above equation (in order to calculate thv' terms -- such as w'thv',
! etc.) for theta_v throughout the model code, such that:
!
! theta_v = theta_l + { (R_v/R_d) - 1 } * thv_ds * r_t
!                   + [ {L_v/(C_p*exner)} - (R_v/R_d) * thv_ds ] * r_c;
!
! where thv_ds is used as a reference value to approximate theta_l.
! However, since only the mean value of theta_v (thvm) is desired in this
! subroutine, and since an accurate calculation is desired due to the fact
! that model pressure, exner, and density rely on this calculation of
! thvm, the exact version is used in this subroutine.

Since pressure, exner, and density are effected, there are some slight changes
to a few cases, but nothing major.

Additionally, in the process of studying this, I looked at the code for
the linearized version of the theta_v equation.  It was:

theta_v = theta_l + { (R_v/R_d) - 1 } * T0 * r_t
                  + [ {L_v/(C_p*exner)} - (R_v/R_d) * T0 ] * r_c;

where T0 is used as a reference value to approximate theta_l.  Now, the constant
T0 (usually 300 K) may be a decent value to approximate theta_l near the surface,
but as theta_l increases higher in the atmosphere, the use of T0 yields an answer
that becomes less and less accurate.  A much better estimate (for use with the
linearized form of the equation) would be base-state theta_v.  Thus, thv_ds now
replaces T0 where the linearized form of the theta_v equation is used, which is
in pdf_closure, compute_length, and the calculation of thvm in advance_clubb_core.
For the Boussinesq form of the code, this change (by itself -- not including the
initial thvm calculation mentioned above) yields an exact binary answer to the
previous version.  However, this change may have effects on the anelastic code.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@4352 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: a3b343c636efe9189e0a21f38d30e73d4f74a43d