Subroutine vertical_avg, which computes the vertical average of various CLUBB fields, is now density-weighted.  In other words, density weighted vertical averages are now being computed for the purposes of hole-filling and statistical output.
One side-effect of this change is that many statistical vertical
averages (for example, wp2_vert_avg, etc.) now appear to be
more inflated, even if there are no changes to the case itself.
This is due to the fact that the vertical averages are now
calculated differently.  In the case of wp2, all the action takes
place in the lower-to-middle parts of the vertical profile.
Of course, density decreases with altitude, and thus is greater
in the lower parts of the profile.  Thus, the levels with an
appreciable value of wp2 are now more highly weighted than they
were before, and the levels with a minimal value of wp2 are now
more lightly weighted than they were before.  Thus, the vertical
average of wp2 increases.  The effect is more pronounced for cases
that have larger vertical domains (the deep convective cases).

In regards the hole-filling, the calculation of vertical averages
are now density-weighted.  However, I have not altered the rest
of the hole-filling code.  In other words, the calculation of how
much mass needes to be deducted is density weighted, but the code
that actually deducts the mass from the rest of the profile still
does so in a manner that is proportional to the amount of 'field',
rather than the amount of density-weighted 'field'.

As of this point, here are the timestep test results:

5 min.:  ARM 97;
10 min.: ARM 97, TWP-ICE;
15 min.: TWP-ICE;
20 min.: ARM 97, TWP-ICE;
30 min.: ARM 97, GABLS 3, TWP-ICE;
45 min.: ARM 97, GABLS 3, RICO, TWP-ICE;
60 min.: ARM 97, cloud feedback s11, GABLS 3, TWP-ICE.

There were 17 total case failures.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@4364 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 337ad5d594c677af00be06af0ad2bb4df54adc10