I came up with what I think is the only possible good solution to the formatting problem with the file hyper_diffusion_4th_ord.F90.
In a nutshell, I took the old "Description" sections (pre-change, 
r4500) and put them with the new code (thank you, "cat" command).
I then placed a note (or a legal disclaimer) on top of the 
"Description" sections that states the following:

! Note:  In the "Description" section of this function, the variable
!        "invrs_dzm" will be written as simply "dzm", and the variable
!        "invrs_dzt" will be written as simply "dzt".  This is being done as
!        as device to save space and to make some parts of the description
!        more readable.  This change does not pertain to the actual code.

The problem is that the descriptive comments need to be (1) easily readable,
giving the reader a quick understanding of what's going on, and (2) fit
within the 100 character limit for the code browser (and personally, I prefer
an 80 character limit, but that isn't always possible).

After analyzing the diagrams, it's just ***not*** possible for the diagrams to 
use "invrs_dzt" and "invrs_dzm", stay readable, and stay under 100 characters.
As stated above, 100 characters is a requirement for the code browser (and
just good coding practice anyway), and if they aren't readable, why have them?
By readable, I mean that the mathematics represented in the terms in the 
diagrams need at least a small degree of indention so that the reader can
discern "what goes with what" -- operators, parentheses, etc.  Without this,
the mathematical terms became one big, jumbled, ugly, disgusting, gut-wrenching,
obscene mess, and would only serve to make life more difficult for the reader.
Already, I had a minimal degree of indention -- I would have preferred more, but
the code browser limited me.  Using "invrs_dzm" and "invrs_dzt" in the diagrams,
as opposed to "dzm" and "dzt", would only force more characters into each row,
as many as 30 more for some rows.  The degree of idention would have to go to 0
in order to fit it all within the 100 character limit, rendering the diagrams
unreadable, in which case we might as well just blow them away.

Thus, the old Descriptions with a binding legal disclaimer is the best way to go.




git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@4615 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 61c9a59d5264b45bdf00f44036c413b127e02e32