For ticket #350.
This adds a lookup table for saturation vapor pressure for water in the saturation module. This is only called within the mixing length code, and is enabled through a l_sat_mixrat_lookup in model_flags.

This provides a speedup of approximately .75-1%.

Also, some of the code in the saturation module has been refactored--specifically, the code that selects between the various SVP approximations has been spun off to seperate functions.


git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@4756 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 403c967ea54a186ddb61ab78c56ea9be938815f4