Removed extraneous arguments that arose due to the changes for ticket #386. I also made the subroutines in advance_xm_wpxp conform to the coding rule about procedure call dummy arguments being on the same lines as the procedure declaration itself. I.e.
call f ( x, y, &
         z )
...

sub  f ( x, y, &
         z )

For reasons that remain a mystery our usual noisy cases change a little.  I'm
starting to suspect it has something to do with g95 or the inclusion of 
optimization in the compiler flags. Perhaps the latter can lead to a subtle 
re-ordering of the math when small changes in subroutine arguments are made.


git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@4982 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 25b11f5d5aa6faa192c4c2450365ad7073b6d0f2