For ticket #396.
Refactors the GMRES solving code for the wp2/wp3 matrices into its own subroutine, wp23_gmres. This is then controlled with the l_gmres flag, rather than having multiple if statements to handle the various calls needed to solve with GMRES.

In addition, this means that no MKL-specific code (that is, anything that needs to be surrounded with #ifdef MKL and #endif /* MKL */) exists in the wp23_solve subroutine--all of that has been moved to wp23_gmres.

CLUBB's output has been tested to be identical before and after this change.


git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@4991 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 6b103602390b38af97318b2d4d729faaa8bd0add