I added model constant parab_cyl_max_input to the model.
The parameter parab_cyl_max_input is the largest magnitude that the input to
the parabolic cylinder function is allowed to have.  When the value of the
input to the parabolic cylinder function is too large in magnitude
(depending on the order of the parabolic cylinder function), overflow
occurs, and the output of the parabolic cylinder function is +/-Inf.  The
parameter parab_cyl_max_input places a limit on the absolute value of the
input to the parabolic cylinder function.  When the value of the potential
input exceeds this parameter (usually due to a very large ratio of ith PDF
component mean of x to ith PDF component standard deviation of x), the
variable x is considered to be constant and a different version of the
equation called.

The largest allowable magnitude of the input to the parabolic cylinder
function (before overflow occurs) is dependent on the order of parabolic
cylinder function.  However, after a lot of testing, it was determined that
an absolute value of 375 works well for an order of 12 or less.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@5107 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 9fd280f68d7d0de63b310d25a8d1307dd5caa191