I have updated the KK microphysics code in a couple of major ways.
First, all of the KK exponents (the exponents that correspond to variables
found in the KK evaporation, autoconversion, accretion, mean volume radius
equations) are declared in parameters_microphys.F90 and have their respective
values assigned in subroutine init_microphys, found in microphysics_driver.F90.
This brings the KK exponents in line with all the other KK and microphysics
parameters.

Secondly, the entire KK code is now compatible with the use of kind core_rknd.
There are no errors or compiler warners generated when the code is compiled
using either single precision or double precisions as kind core_rknd.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@5649 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 938f50e6c23fc78eb30f6cf786fe281f4286dc48