As mentioned in r 5651, I have found out that my decision to specify correlations in terms of r_t and theta_l, rather than in terms of extended liquid water mixing ratio, s, was not a good one. For starters, we need specified correlations for both s and t for the covariance code. We need to specify the same number of correlations whether we use r_t and theta_l or s and t. If you know the correlations in terms of s and t, there are equations to find the correlations in terms of r_t and theta_l, and likewise if you know the correlations in terms of r_t and theta_l, there are equations to find the correlations in terms of s and t.
The remainder of the CLUBB model is already set up to use s and t (for example,
Latin Hypercube) in regards to PDF parameters.

I am undoing my changes, committed on April 26, 2011, that changed the KK code
away form using specified correlations in terms of s, and instead started using
specified correlations in terms of r_t and theta_l, and then used an equation to
find the correlations in terms of s. The original change only made slight
differences on the nightly plots, anyway.

This change takes care of the rest of the original change.  This revision
undoes r5147 and r5148.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@5652 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: b315a0a0ac44c83e4c8c1794e3b08f8e9711a7c7