I added statistical output for the nine covariances involving w, r_t, or theta_l and a KK microphysical tendency.  They are:  w_KK_evap_covar, rt_KK_evap_covar, thl_KK_evap_covar, w_KK_auto_covar, rt_KK_auto_covar, thl_KK_auto_covar, w_KK_accr_covar, rt_KK_accr_covar, and thl_KK_auto_covar.
Since all the variables used in the calculation of these values are located on
thermodynamic grid levels, these covariances are located on thermodynamic grid
levels.  Later, these covariances are used in different combinations to
calculate microphysics tendency terms for model variances and covariances
(ex.: rtp2 microphysics tendency).  The values of those tendencies will need
to be interpolated to momentum levels, and the statistics for those tendency
variables will be stored in the momentum level file.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@5782 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: bc3c3ca1f73709002ae36cb2756758942606920a