For clubb:ticket:602. I finished the generalization of setup_clubb_pdf_params.F90. The code should now be able to deal with an arbitrary set of hydrometeors, as long as the corresponding indices (e.g. for the hydrometeor rrain: iiPDF_rrain, iirrainm, iirr) are set up properly. The code is assuming that the iiPDF indices are setup in the following order: s, t, w, Ncn, hydrometeors (increasing from left to right) For the loops in which all the hydrometeor means/stdevs/correlations are computed, I implemented a map from the iiPDF indices to the ii indices that are used to address in-precip hydrometeors in setup_clubb_pdf_params.F90, namely {{{hm_ip_idx}}}.
I merged my local copy with r6516. After the merge I obtained bit-by-bit identical results for the cases atex, dycoms2_rf02_ds, fire and rico. For the cases arm_97, cgils_s6 and mpace_b_LH the results differed in the correlation values, but everything else was identical. The difference in the correlations was arising through a bug in the repository version. For details see clubb:comment:11:ticket:602. 


git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@6517 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 7acbb0df56fd36c24bf760ccfb88190e268f787f