I am updating the surface variance scheme based on Andre et al. 1978 (turned off by default in CLUBB).  The paper contains an equation for convective velocity, w*:
w* = ( ( 1 / T0 ) * g * <w'thl'>|_sfc * z_i )^(1/3)

where z_i is the height of the mixing layer.  We were erroneously using a
constant value of 1 m. for this height.  It should be something like Lscale at
the surface.  In this revision, I am making the change.  This should boost the
magnitude of w* by a factor of Lscale**(1/3).  The value of w* plays directly
into the equations for up2_sfc and vp2_sfc.  This should boost the value of up2
and vp2 at the surface when wpthlp_sfc is positive.

I also did some cleaning up of the file surface_varnce_module.F90.

The GrADS output results from this revision when l_andre_1978 remains turned off
(which it is by default) are an exact match to the GrADS output results from the
previous revision.

Obligatory Larson Group Internal Trac System Reference: clubb:ticket:230.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@6870 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: c83985fe3efffdead3d4b2c18f0d38e92d1dd5a8