Previously, I noticed a problem with the CLUBB code that calculates Ncnm from Nc_in_cloud (and also Nc_in_cloud from Ncnm). The problem is that these equations required cloud fraction as calculated by the PDF. However, after cloud fraction is originally calculated from the PDF in pdf_closure, it undergoes alterations in a few spots (trapezoidal rule, etc.). This makes cloud fraction inconsistent with the PDF. As long as cloud_frac1 and cloud_frac2 (PDF component cloud fractions) are not altered, cloud_frac can simply be recalculated locally from these variables.
However, some codes were found that are turned off by default that would make
cloud_frac1 and cloud_frac2 inconsistent with the PDF if they are turned on.
There needs to be some considerable refactoring done.  For now, I have placed
a comment, a write statement, and a stop statement in each section of code that
has this problem.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@6876 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: bc8d1f03376ccf1d40feac4b6cedeea724ad381d