I am making the code that sets PDF array indices more robust.
In r6907, I committed new code with the following message:

Not too long ago, I ran across a problem in CLUBB. When looping over all the
hydrometeors in hydromet_dim and mapping the hydromet index to the PDF index, I
ran across an array out of bounds error. In a nutshell, it turned out that the
hydrometeor set of indices (iirrainm, iiNrm, etc.) were not entirely a subset of
the PDF set of indices (iiPDF_rrain, iiPDF_Nr, etc.). To make a point, each
index in the hydrometeor set of indices should also correspond to an index in
the PDF set of indices. In other words, the hydrometeor set of indices should be
a proper subset of the PDF set of indices. When there are indices found in the
hydrometeor set that are not found in the PDF set, this causes problems.
Anyway, this problem was occuring when Morrison microphysics was run with
l_graupel turned off. The hydrometeor indices were not set up properly for this
specific scenario, whereas the PDF set of indices were set up properly to handle
this.

CLUBB r6907 solved the problem for Morrison microphysics.  However, today, when
I tried to map hydrometeor indices to PDF indices, there was an array out of
bound error for COAMPS microphysics.

Unlike Morrison microphysics, COAMPS microphysics does not use the fancy flags
l_ice_microphys and l_graupel.

Instead, the solution here is to make subroutine setup_pdf_indices more robust
-- able to handle anythign that is thrown at it.

I have done that in this revision.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@6910 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 6873eb6f154a956445dac3363bc37c915b48534e