I am adding new stats corr_st_1_ca and corr_st_2_ca.
The true correlations between s and t in each PDF component are calculated
by an equation and stored as part of CLUBB's PDf parameters.  However, the
CLUBB code has an option, l_fix_s_t_correlations, to set the component
correlations between s and t to a constant, prescribed value.  This is mainly
used for SILHS (pathetic).  The statistical outputs "corr_st_1" and "corr_st_2"
contain the values of the component correlation between s and t as calculated
by the equation as part of PDF parameters.  We should also output whatever
value is used in the correlation arrays, whether or not it matches the value
calculated by the equation.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@6919 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 9b4102f95f321208dad2d84ba2ec28ce102b25d1