I am enhancing run_bindiff_all.bash to be able to compare netCDF files.  It can now check for binary differences in both GrADS and netCDF output files.  The usage is the same as before.
Along the way, I found a problem -- netCDF files from two runs that are exactly
the same would still produce binary differences.  I found that the time that
the runs started (technically the time the netCDF output file was created) was
being written as a global variable to the netCDF output file.  For example,
if a ran the ARM case, made no changes, and then ran the ARM case again 1 minute
later, the netCDF output files from the two runs would contain binary differences.
I found the line in CLUBB's output_netcdf.F90 where the "created_on" global variable
was written.  I then added a flag to CLUBB, l_output_file_run_date, that can be
enabled if the user wants this global variable, and then I set it to .false. by
default.  This doesn't seem to be very important anyway -- and if the user wants
to be the date and time of the file, he/she can just do a ls -la.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@6931 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 79d2db780ff774dcc81480e3fbcbe7507ef37cbd