For clubb:ticket:610, I am handling the `l_const_Nc_in_cloud` code consistently in `setup_clubb_pdf_params.F90`. For one, the variance of `Ncn` is now zero when `l_const_Nc_in_cloud = .true.`.
As a result of this change, SILHS will now correctly output a "delta" function for `Ncn` when `l_const_Nc_in_cloud = .true.`, since I have set the variance of `Ncn` to be zero.

Output is bin-diff identical for cases that do not use SILHS. When SILHS is used and `l_const_Nc_in_cloud = .false.`, output is also bin-diff identical. Unfortunately, when `l_const_Nc_in_cloud = .true.` and SILHS is used, the output changes. I think this is a necessary artifact of the sampling process. Since a correlation between a delta and a normal/lognormal variate is undefined, I set the correlations between `Ncn` and the other variates to zero. Before this commit, these correlations were not zero, so all the other variates were correlated with a lognormal `Ncn` distribution that was incorrect and was eventually thrown out anyway. Now, `Ncn` is not correlated with any other variate when `l_const_Nc_in_cloud = .true.`. However, I don't think this will affect the convergence characteristics of SILHS, since it doesn't make any meaningful change to the PDF.

Hopefully...


git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@7130 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 268cd01c0b83981ec54eba612d93b825e7286504