After a series of suspenseful derivations on the chalkboard, I finally have my assertion check working!
This is an assertion check in SILHS that makes sure that the values of `chi_i`, `stdev_chi_i`, and `cloud_frac_i` are consistent with each other in both PDF components at every height level. This assertion check was desired because of a recent trigger of another assertion check (in SILHS) about half of all sample points not being in cloud (see clubb:ticket:720 for a discussion on this).

I'm not really sure how well it works. I hacked away at it until all my evil test cases caused the assertion check to trigger, but I would still hesitate to call it "super robust" until I've done some more testing.

Anyway, output is bin-diff identical for all the SILHS cases (they all pass the assertion check on my computer with `gfortran`).


git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@7145 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: d974f62b555fce06c50aceb42bcd2c88d9256bfe