I am updating some code in regards to the CLUBB tuner.
I eventually figured out how to set up and run the CLUBB tuner for the RICO
case.  Along the way, I encountered some difficulties.  After the tuner had
finished tuning, it did not rerun the case to produce the new results.  The
issue had to do with the new variable "thlp2_rad_cloud_frac_thresh".  It was
being truncated after "thlp2_rad_cloud_".  This was causing a runtime error.
This is bizarre because the tuner looks to be working fine for the FIRE case
on the nightly tests.

Anyway, it runs fine now for my RICO case.

Larson Group Internal Trac System Reference: clubb:ticket:735.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@7252 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 246326eaf2de396329688ce138aa5f35bdb6aa9c