For clubb:ticket:736, I am generalizing the check that half of all sample points are in cloud in normal space.
The assertion check will now ensure that the cloudiness of each sample point is consistent in both spaces. That is, it will assert that every sample point is either cloudy in both uniform and normal space, or is clear in both uniform and normal space. In theory, we could run the new assertion check at every vertical level, but to save computational expense (even though this is a `debug_level(2)` check), I am only doing it at `k_lh_start`.

This modification should make the assertion check more robust. Also, combined with the check for half cloudy points in uniform space, it should still ensure that half of all sample points are in cloud in normal space. It can also easily be generalized to account for the fact that half of all sample points might not be in cloud (for example, in the new sampling scheme).

The `arm_97`, `rico_lh`, `twp_ice`, and `mpace_b_lh` cases were tested and found to pass the new assertion check.


git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@7276 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 4fb6611e9affbccd4df4e9922decee0073cebdd7