I changed all the iwprrp, iwpNrp, etc., statistical indices to an array, similar to what I have done for ithlphmp, iwp2hmp, irtphmp, and all hydrometeor PDF statistics.  This has the advantage of only producing the necessary statistics for the microphysics scheme that was used.  It gets rid of a lot of wasted statistics.
The GrADS output files from this revision are an exact match to the GrADS output
files from the previous revision for all zt and sfc files.  The zm files are
different because of the removed/reorder stats.

Continuing with our series ...

The "that one guy" Memorial Trac System Reference: clubb:ticket:639, and
somewhat related to clubb:ticket:730.

This is the Memorial Trac System Reference to that one guy that was in the old
EMS E468 room in late 2003/early 2004.  That one guy that was kind of a
system administrator type guy (from UWM ITS, I think).  That one guy that had
dark hair.  He wasn't around that long, and I have absolutely no recollection of
his name -- but here's to him!



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@7293 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 24a3731da57d1c1acbb08efc8911c2839d4fdedf