I don't need to reset true correlations involving Ncn to 0 when l_const_Nc_in_cloud is used (which means Ncnp2_on_Ncnm2 = 0).  When this happens, correlations involving Ncn are technically undefined.  However, as Ncnp2_on_Ncnm2 approaches 0, corr_x_Ncn_i approaches corr_x_Ncn_i_n.  The true correlations involving Ncn can simply be set to their normal space equivalents.  Since l_const_Nc_in_cloud is used, none of this matters in the results anyway.  It just makes it easier to compare the output of the new correlation arrays to their previous versions (before normal space correlations were specified).
The results in GrADS output files from this revision are different than the
results in GrADS output files from the previous revision for some case and for
zt output files only.  This happens because the corr_x_Ncn_i is no longer set
to 0.  Otherwise, other output files are not affected and case results are not
affected by this change.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@7591 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 6b9a95785d0060c0d7ff36e43e290ac091ac8867