In setup_pdf_parameters, hydrometp2_zt is calculated and used locally.  Then, in a subroutine called by setup_pdf_parameters (the mu, sigma code subroutine), its value is calculated again.  There isn't any reason to do things like that, the value can just be passed through.  So, in this revision, I am feeding the value of hydrometp2_zt through to the subroutine that calculates mu and sigma values.
The formulas to calculate hydrometp2 are mathematically equivalent between the
two codes (the "main" one and the one that we used to use to recalculate with).
However, they are not structured exactly the same, so results are numerically
different at the roundoff level.  This doesn't bother most, but ARM 97 freaks
out.

Trac System Reference: clubb:ticket:755.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@7597 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 1824e13cf70d7252fd4a9d635a3f0cb84c887d44