For clubb:ticket:775:
I dug up some old timing code from an obsolete CLUBB branch and found that it gives accurate results. Therefore, I am committing it to the trunk of CLUBB, inside the `CLUBB_core` directory. It is not used right now; it is simply designed to be a utility for code analysis.

I cleaned up the code and added comments to make it suitable for the "official" CLUBB.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@7934 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 10dcb3e8aaf7b5d36530bcbbb34e3787686f6403