For cam:ticket:52:
I moved `l_stability_correct_tau_zm` to `model_flags.F90`, and made it a variable instead of a compile-time constant. This should be the only change we need to make to CLUBB to do away with the branch.

CLUBB standalone output is bin-diff identical for a fine selection of CLUBB cases.


git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@7978 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 25a8690cba226668f8eb62e75b86b6a43782a8c2